+++ source/databricks/tests/test_time_series_transformer.py
@@ -25,7 +25,7 @@ from pyspark.sql.types import (
 )
 from datetime import datetime
 from decimal import Decimal
-from package.transforms.jsonTransformer import (
+from package.transforms.time_series_transformer import (
     transform_unprocessed_time_series_to_points,
 )
